+++ src/Test/Demo.php
@@ -71,20 +71,20 @@ class Demo extends TestCommon
 
         $rule = [
             "id" => "required|/^\d+$/",
-            "name" => "required|string|len<=>[8,32]",
-            "gender" => "required|(s)[male,female]",
+            "name" => "required|string|length><=[8,32]",
+            "gender" => "required|<string>[male,female]",
             "dob" => "required|dob",
             "age" => "required|check_age[@gender,30] >> @this is wrong",
             "height[or]" => [
-                "required|=(@height_unit,cm)|<=>=[100,200] >> @this should be in [100,200] when height_unit is cm",
-                "required|=(@height_unit,m)|<=>=[1,2] >> @this should be in [1,2] when height_unit is m",
+                "required|=(@height_unit,cm)|>=<=[100,200] >> @this should be in [100,200] when height_unit is cm",
+                "required|=(@height_unit,m)|>=<=[1,2] >> @this should be in [1,2] when height_unit is m",
             ],
-            "height_unit" => "required|(s)[cm,m]",
+            "height_unit" => "required|<string>[cm,m]",
             "weight[or]" => [
-                "required|=(@weight_unit,kg)|<=>=[40,100]",
-                "required|=(@weight_unit,lb)|<=>[88,220]",
+                "required|=(@weight_unit,kg)|>=<=[40,100]",
+                "required|=(@weight_unit,lb)|><=[88,220]",
             ],
-            "weight_unit" => "required|(s)[kg,lb]",
+            "weight_unit" => "required|<string>[kg,lb]",
             "email" => "required|email",
             "phone" => "required|/(^1[3|4|5|6|7|8|9]\d{9}$)|(^09\d{8}$)/ >> phone number error",
             "ip" => "optional|ip",
@@ -92,23 +92,23 @@ class Demo extends TestCommon
             "education" => [
                 "primary_school" => "required|=[Qiankeng Xiaoxue]",
                 "junior_middle_school" => "required|!=[Foshan Zhongxue]",
-                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[8]",
-                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[8]",
+                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[8]",
+                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[8]",
             ],
             "company" => [
-                "name" => "required|len<=>[8,64]",
+                "name" => "required|length><=[8,64]",
                 "website" => "required|url",
-                "country" => "optional|len<=[32]",
-                "addr" => "required|len>[16]",
-                "postcode" => "optional|len<[16]|check_postcode(@parent)",
+                "country" => "optional|length<=[32]",
+                "addr" => "required|length>[16]",
+                "postcode" => "optional|length<[16]|check_postcode(@parent)",
                 "colleagues.*" => [
-                    "name" => "required|string|len<=>[3,32]",
-                    "position" => "required|(s)[Reception,Financial,PHP,JAVA]"
+                    "name" => "required|string|length><=[3,32]",
+                    "position" => "required|<string>[Reception,Financial,PHP,JAVA]"
                 ],
                 "boss" => [
                     "required|=[Mike]",
-                    "required|(s)[Johnny,David]",
-                    "optional|(s)[Johnny,David]"
+                    "required|<string>[Johnny,David]",
+                    "optional|<string>[Johnny,David]"
                 ]
             ],
             "favourite_food[optional].*" => [
@@ -184,24 +184,24 @@ class Demo extends TestCommon
 
         $rule = [
             "id" => "required|/^\d+$/",
-            "name" => "required|string|len<=>[8,32]",
-            "gender" => "required|(s)[male,female]",
+            "name" => "required|string|length><=[8,32]",
+            "gender" => "required|<string>[male,female]",
             "dob" => "required|dob",
             "age" => "required|check_age[@gender,30] >> @this is wrong",
             "height" => [
                 "[or]" => [
-                    "required|=(@height_unit,cm)|<=>=[100,200] >> @this should be in [100,200] when height_unit is cm",
-                    "required|=(@height_unit,m)|<=>=[1,2] >> @this should be in [1,2] when height_unit is m",
+                    "required|=(@height_unit,cm)|>=<=[100,200] >> @this should be in [100,200] when height_unit is cm",
+                    "required|=(@height_unit,m)|>=<=[1,2] >> @this should be in [1,2] when height_unit is m",
                 ]
             ],
-            "height_unit" => "required|(s)[cm,m]",
+            "height_unit" => "required|<string>[cm,m]",
             "weight" => [
                 "[or]" => [
-                    "required|=(@weight_unit,kg)|<=>=[40,100]",
-                    "required|=(@weight_unit,lb)|<=>[88,220]",
+                    "required|=(@weight_unit,kg)|>=<=[40,100]",
+                    "required|=(@weight_unit,lb)|><=[88,220]",
                 ]
             ],
-            "weight_unit" => "required|(s)[kg,lb]",
+            "weight_unit" => "required|<string>[kg,lb]",
             "email" => "required|email",
             "phone" => "required|/(^1[3|4|5|6|7|8|9]\d{9}$)|(^09\d{8}$)/ >> phone number error",
             "ip" => "optional|ip",
@@ -209,25 +209,25 @@ class Demo extends TestCommon
             "education" => [
                 "primary_school" => "required|=[Qiankeng Xiaoxue]",
                 "junior_middle_school" => "required|!=[Foshan Zhongxue]",
-                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[8]",
-                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[8]",
+                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[8]",
+                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[8]",
             ],
             "company" => [
-                "name" => "required|len<=>[8,64]",
+                "name" => "required|length><=[8,64]",
                 "website" => "required|url",
-                "country" => "optional|len<=[32]",
-                "addr" => "required|len>[16]",
-                "postcode" => "optional|len<[16]|check_postcode(@parent)",
+                "country" => "optional|length<=[32]",
+                "addr" => "required|length>[16]",
+                "postcode" => "optional|length<[16]|check_postcode(@parent)",
                 "colleagues" => [
                     ".*" => [
-                        "name" => "required|string|len<=>[3,32]",
-                        "position" => "required|(s)[Reception,Financial,PHP,JAVA]"
+                        "name" => "required|string|length><=[3,32]",
+                        "position" => "required|<string>[Reception,Financial,PHP,JAVA]"
                     ]
                 ],
                 "boss" => [
                     "required|=[Mike]",
-                    "required|(s)[Johnny,David]",
-                    "optional|(s)[Johnny,David]"
+                    "required|<string>[Johnny,David]",
+                    "optional|<string>[Johnny,David]"
                 ]
             ],
             "favourite_food" => [
@@ -315,20 +315,20 @@ class Demo extends TestCommon
 
         $rule = [
             "id" => 'required|/^\d+$/ >> { "required": "Users define - @this is required", "preg": "Users define - @this should be \"MATCHED\" @preg"}',
-            "name" => "required|string|len<=>[8,32]",
-            "gender" => "required|(s)[male,female]",
+            "name" => "required|string|length><=[8,32]",
+            "gender" => "required|<string>[male,female]",
             "dob" => "required|dob",
             "age" => "required|check_age[@gender,30] >> @this is wrong",
             "height[or]" => [
-                "required|=(@height_unit,cm)|<=>=[100,200] >> @this should be in [100,200] when height_unit is cm",
-                "required|=(@height_unit,m)|<=>=[1,2] >> @this should be in [1,2] when height_unit is m",
+                "required|=(@height_unit,cm)|>=<=[100,200] >> @this should be in [100,200] when height_unit is cm",
+                "required|=(@height_unit,m)|>=<=[1,2] >> @this should be in [1,2] when height_unit is m",
             ],
-            "height_unit" => "required|(s)[cm,m]",
+            "height_unit" => "required|<string>[cm,m]",
             "weight[or]" => [
-                "required|=(@weight_unit,kg)|<=>=[40,100] >> @this should be in [40,100] when height_unit is kg",
-                "required|=(@weight_unit,lb)|<=>[88,220] >> @this should be in [88,220] when height_unit is lb",
+                "required|=(@weight_unit,kg)|>=<=[40,100] >> @this should be in [40,100] when height_unit is kg",
+                "required|=(@weight_unit,lb)|><=[88,220] >> @this should be in [88,220] when height_unit is lb",
             ],
-            "weight_unit" => "required|(s)[kg,lb]",
+            "weight_unit" => "required|<string>[kg,lb]",
             "email" => "required|email",
             "phone" => "required|/(^1[3|4|5|6|7|8|9]\d{9}$)|(^09\d{8}$)/ >> phone number error",
             "ip" => "optional|ip",
@@ -336,23 +336,23 @@ class Demo extends TestCommon
             "education" => [
                 "primary_school" => "required|=[Qiankeng Xiaoxue]",
                 "junior_middle_school" => "required|!=[Foshan Zhongxue]",
-                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[18]",
-                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[8]",
+                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[18]",
+                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[8]",
             ],
             "company" => [
-                "name" => "required|len<=>[8,64]",
+                "name" => "required|length><=[8,64]",
                 "website" => "required|url",
-                "country" => "optional|len>=[6]",
-                "addr" => "required|len>[16]",
-                "postcode" => "optional|len<[16]|check_postcode(@parent)",
+                "country" => "optional|length>=[6]",
+                "addr" => "required|length>[16]",
+                "postcode" => "optional|length<[16]|check_postcode(@parent)",
                 "colleagues.*" => [
-                    "name" => "required|string|len<=>[3,32]",
-                    "position" => "required|(s)[Reception,Financial,PHP,JAVA]"
+                    "name" => "required|string|length><=[3,32]",
+                    "position" => "required|<string>[Reception,Financial,PHP,JAVA]"
                 ],
                 "boss" => [
                     "required|=[Mike]",
-                    "required|(s)[Johnny,David]",
-                    "optional|(s)[Johnny,David]"
+                    "required|<string>[Johnny,David]",
+                    "optional|<string>[Johnny,David]"
                 ]
             ],
             "favourite_food[optional].*" => [
@@ -438,24 +438,24 @@ class Demo extends TestCommon
 
         $rule = [
             "id" => "required|int|/^\d+$/",
-            "name" => "required|string|len<=>[8,32]",
-            "gender" => "required|(s)[male,female]",
+            "name" => "required|string|length><=[8,32]",
+            "gender" => "required|<string>[male,female]",
             "dob" => "required|dob",
             "age" => "required|check_age[@gender,30] >> @this is wrong",
             "height" => [
                 "[or]" => [
-                    "required|=(@height_unit,cm)|<=>=[100,200] >> @this should be in [100,200] when height_unit is cm",
-                    "required|=(@height_unit,m)|<=>=[1,2] >> @this should be in [1,2] when height_unit is m",
+                    "required|=(@height_unit,cm)|>=<=[100,200] >> @this should be in [100,200] when height_unit is cm",
+                    "required|=(@height_unit,m)|>=<=[1,2] >> @this should be in [1,2] when height_unit is m",
                 ]
             ],
-            "height_unit" => "required|(s)[cm,m]",
+            "height_unit" => "required|<string>[cm,m]",
             "weight[or]" => [
                 "[or]" => [
-                    "required|=(@weight_unit,kg)|<=>=[40,100] >> @this should be in [40,100] when height_unit is kg",
-                    "required|=(@weight_unit,lb)|<=>[88,220] >> @this should be in [88,220] when height_unit is lb",
+                    "required|=(@weight_unit,kg)|>=<=[40,100] >> @this should be in [40,100] when height_unit is kg",
+                    "required|=(@weight_unit,lb)|><=[88,220] >> @this should be in [88,220] when height_unit is lb",
                 ]
             ],
-            "weight_unit" => "required|(s)[kg,lb]",
+            "weight_unit" => "required|<string>[kg,lb]",
             "email" => "required|email",
             "phone" => "required|/(^1[3|4|5|6|7|8|9]\d{9}$)|(^09\d{8}$)/ >> phone number error",
             "ip" => "optional|ip",
@@ -463,25 +463,25 @@ class Demo extends TestCommon
             "education" => [
                 "primary_school" => "required|=[Qiankeng Xiaoxue]",
                 "junior_middle_school" => "required|!=[Foshan Zhongxue]",
-                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[18]",
-                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|len>[8]",
+                "high_school" => "if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[18]",
+                "university" => "!if(=(@junior_middle_school,Mianhu Zhongxue))|required|length>[8]",
             ],
             "company" => [
-                "name" => "required|len<=>[8,64]",
+                "name" => "required|length><=[8,64]",
                 "website" => "required|url",
-                "country" => "optional|len>=[6]",
-                "addr" => "required|len>[16]",
-                "postcode" => "optional|len<[16]|check_postcode(@parent)",
+                "country" => "optional|length>=[6]",
+                "addr" => "required|length>[16]",
+                "postcode" => "optional|length<[16]|check_postcode(@parent)",
                 "colleagues" => [
                     ".*" => [
-                        "name" => "required|string|len<=>[3,32]",
-                        "position" => "required|(s)[Reception,Financial,PHP,JAVA]"
+                        "name" => "required|string|length><=[3,32]",
+                        "position" => "required|<string>[Reception,Financial,PHP,JAVA]"
                     ]
                 ],
                 "boss" => [
                     "required|=[Mike]",
-                    "required|(s)[Johnny,David]",
-                    "optional|(s)[Johnny,David]"
+                    "required|<string>[Johnny,David]",
+                    "optional|<string>[Johnny,David]"
                 ]
             ],
             "favourite_food" => [
@@ -559,7 +559,7 @@ class Demo extends TestCommon
             // id 是必要的且必须匹配正则 /^\d+$/， >> 后面的required 和 正则对应的报错信息
             "id" => 'required|/^\d+$/ >> { "required": "用户自定义 - @this 是必要的", "preg": "用户自定义 - @this 必须匹配 @preg" }',
             // name 是必要的且必须是字符串且长度在区间 【8，32)
-            "name" => "required|string|len<=>[8,32]",
+            "name" => "required|string|length><=[8,32]",
             "email" => "required|email",
             "phone" => "required|/(^1[3|4|5|6|7|8|9]\d{9}$)|(^09\d{8}$)/ >> 用户自定义 - phone number 错误",
             // ip 是可选的
@@ -572,18 +572,18 @@ class Demo extends TestCommon
                 "university" => "optional|string",
             ],
             "company" => [
-                "name" => "required|len<=>[8,64]",
+                "name" => "required|length><=[8,64]",
                 "website" => "required|url",
                 "colleagues.*" => [
-                    "name" => "required|string|len<=>[3,32]",
+                    "name" => "required|string|length><=[3,32]",
                     // company.colleagues.*.position 必须等于 Reception,Financial,PHP,JAVA 其中之一
-                    "position" => "required|(s)[Reception,Financial,PHP,JAVA]"
+                    "position" => "required|<string>[Reception,Financial,PHP,JAVA]"
                 ],
                 // 以下三个规则只对 boss.0, boss.1, boss.2 有效，boss.3 及其他都无效 
                 "boss" => [
                     "required|=[Mike]",
-                    "required|(s)[Johnny,David]",
-                    "optional|(s)[Johnny,David]"
+                    "required|<string>[Johnny,David]",
+                    "optional|<string>[Johnny,David]"
                 ]
             ],
             // favourite_food 是可选的索引数组，允许为空