+++ src/Test/Rule/TestRuleDefault.php
@@ -93,8 +93,8 @@ trait TestRuleDefault
             "key_float_str" => "optional|==[\"1.1\"]|strictly_equal['1.1']",
             "key_bool_true" => "optional|==[TRUE]|strictly_equal[true]",
             "key_bool_false" => "optional|==[FALSE]|strictly_equal[false]",
-            "key_bool_str_true" => "optional|==[\"true\"]|strictly_equal['true']",
-            "key_bool_str_FALSE" => "optional|==[\"FALSE\"]|strictly_equal['FALSE']",
+            "key_bool_string_true" => "optional|==[\"true\"]|strictly_equal['true']",
+            "key_bool_string_FALSE" => "optional|==[\"FALSE\"]|strictly_equal['FALSE']",
         ];
 
         $cases = [
@@ -106,8 +106,8 @@ trait TestRuleDefault
                     "key_float_str" => "1.1",
                     "key_bool_true" => true,
                     "key_bool_false" => FALSE,
-                    "key_bool_str_true" => "true",
-                    "key_bool_str_FALSE" => "FALSE",
+                    "key_bool_string_true" => "true",
+                    "key_bool_string_FALSE" => "FALSE",
                 ]
             ],
             "Valid_data_2" => [
@@ -118,8 +118,8 @@ trait TestRuleDefault
                     "key_float_str" => "1.1",
                     "key_bool_true" => TRUE,
                     "key_bool_false" => FALSE,
-                    "key_bool_str_true" => "true",
-                    "key_bool_str_FALSE" => "FALSE",
+                    "key_bool_string_true" => "true",
+                    "key_bool_string_FALSE" => "FALSE",
                 ]
             ],
             "Invalid_key_int" => [
@@ -194,29 +194,29 @@ trait TestRuleDefault
                 ],
                 "expected_msg" => ["key_bool_false" => "key_bool_false must be strictly equal to bool(false)"]
             ],
-            "Invalid_key_bool_str_true" => [
+            "Invalid_key_bool_string_true" => [
                 "data" => [
-                    "key_bool_str_true" => true,
+                    "key_bool_string_true" => true,
                 ],
-                "expected_msg" => ["key_bool_str_true" => "key_bool_str_true must be strictly equal to string(true)"]
+                "expected_msg" => ["key_bool_string_true" => "key_bool_string_true must be strictly equal to string(true)"]
             ],
-            "Invalid_key_bool_str_true_1" => [
+            "Invalid_key_bool_string_true_1" => [
                 "data" => [
-                    "key_bool_str_true" => "TRUE",
+                    "key_bool_string_true" => "TRUE",
                 ],
-                "expected_msg" => ["key_bool_str_true" => "key_bool_str_true must be strictly equal to string(true)"]
+                "expected_msg" => ["key_bool_string_true" => "key_bool_string_true must be strictly equal to string(true)"]
             ],
-            "Invalid_key_bool_str_FALSE" => [
+            "Invalid_key_bool_string_FALSE" => [
                 "data" => [
-                    "key_bool_str_FALSE" => false,
+                    "key_bool_string_FALSE" => false,
                 ],
-                "expected_msg" => ["key_bool_str_FALSE" => "key_bool_str_FALSE must be strictly equal to string(FALSE)"]
+                "expected_msg" => ["key_bool_string_FALSE" => "key_bool_string_FALSE must be strictly equal to string(FALSE)"]
             ],
-            "Invalid_key_bool_str_FALSE_1" => [
+            "Invalid_key_bool_string_FALSE_1" => [
                 "data" => [
-                    "key_bool_str_FALSE" => "false",
+                    "key_bool_string_FALSE" => "false",
                 ],
-                "expected_msg" => ["key_bool_str_FALSE" => "key_bool_str_FALSE must be strictly equal to string(FALSE)"]
+                "expected_msg" => ["key_bool_string_FALSE" => "key_bool_string_FALSE must be strictly equal to string(FALSE)"]
             ],
         ];
 
@@ -240,8 +240,8 @@ trait TestRuleDefault
             "key_float_str" => "optional|!==[\"1.1\"]|not_strictly_equal['1.1']",
             "key_bool_true" => "optional|!==[TRUE]|not_strictly_equal[true]",
             "key_bool_false" => "optional|!==[FALSE]|not_strictly_equal[false]",
-            "key_bool_str_true" => "optional|!==[\"true\"]|not_strictly_equal['true']",
-            "key_bool_str_FALSE" => "optional|!==[\"FALSE\"]|not_strictly_equal['FALSE']",
+            "key_bool_string_true" => "optional|!==[\"true\"]|not_strictly_equal['true']",
+            "key_bool_string_FALSE" => "optional|!==[\"FALSE\"]|not_strictly_equal['FALSE']",
         ];
 
         $cases = [
@@ -253,8 +253,8 @@ trait TestRuleDefault
                     "key_float_str" => 1.1,
                     "key_bool_true" => "true",
                     "key_bool_false" => "FALSE",
-                    "key_bool_str_true" => true,
-                    "key_bool_str_FALSE" => FALSE,
+                    "key_bool_string_true" => true,
+                    "key_bool_string_FALSE" => FALSE,
                 ]
             ],
             "Valid_data_2" => [
@@ -265,8 +265,8 @@ trait TestRuleDefault
                     "key_float_str" => 1.10,
                     "key_bool_true" => "TRUE",
                     "key_bool_false" => 0,
-                    "key_bool_str_true" => 1,
-                    "key_bool_str_FALSE" => "",
+                    "key_bool_string_true" => 1,
+                    "key_bool_string_FALSE" => "",
                 ]
             ],
         ];
@@ -495,7 +495,7 @@ trait TestRuleDefault
     protected function test_method_between()
     {
         $rule = [
-            "id" => "<>[1,10]|between[1,10]",
+            "id" => "><[1,10]|between[1,10]",
         ];
 
         $cases = [
@@ -538,7 +538,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "<>",
+            "error_tag" => "><",
             "field_path" => "id",
             "parameters" => [
                 "1",
@@ -556,7 +556,7 @@ trait TestRuleDefault
     protected function test_method_greater_lessequal()
     {
         $rule = [
-            "id" => "<=>[1,10]|greater_lessequal[1,10]",
+            "id" => "><=[1,10]|greater_lessequal[1,10]",
         ];
 
         $cases = [
@@ -604,7 +604,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "<=>",
+            "error_tag" => "><=",
             "field_path" => "id",
             "parameters" => [
                 "1",
@@ -622,7 +622,7 @@ trait TestRuleDefault
     protected function test_method_greaterequal_less()
     {
         $rule = [
-            "id" => "<>=[1,10]|greaterequal_less[1,10]",
+            "id" => ">=<[1,10]|greaterequal_less[1,10]",
         ];
 
         $cases = [
@@ -670,7 +670,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "<>=",
+            "error_tag" => ">=<",
             "field_path" => "id",
             "parameters" => [
                 "1",
@@ -688,7 +688,7 @@ trait TestRuleDefault
     protected function test_method_greaterequal_lessequal()
     {
         $rule = [
-            "id" => "<=>=[1,10]|greaterequal_lessequal[1,10]",
+            "id" => ">=<=[1,10]|greaterequal_lessequal[1,10]",
         ];
 
         $cases = [
@@ -736,7 +736,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "<=>=",
+            "error_tag" => ">=<=",
             "field_path" => "id",
             "parameters" => [
                 "1",
@@ -754,7 +754,7 @@ trait TestRuleDefault
     protected function test_method_in_number_array()
     {
         $rule = [
-            "id" => "(n)[1,2,3]|in_number_array[1,2]",
+            "id" => "<number>[1,2,3]|in_number_array[1,2]",
         ];
 
         $cases = [
@@ -809,7 +809,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "(n)",
+            "error_tag" => "<number>",
             "field_path" => "id",
             "parameters" => [
                 "1,2,3"
@@ -826,7 +826,7 @@ trait TestRuleDefault
     protected function test_method_not_in_number_array()
     {
         $rule = [
-            "id" => "!(n)[1,2]|not_in_number_array[1,2,3]",
+            "id" => "!<number>[1,2]|not_in_number_array[1,2,3]",
         ];
 
         $cases = [
@@ -894,7 +894,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "!(n)",
+            "error_tag" => "!<number>",
             "field_path" => "id",
             "parameters" => [
                 "1,2,3"
@@ -911,7 +911,7 @@ trait TestRuleDefault
     protected function test_method_in_string_array()
     {
         $rule = [
-            "id" => "(s)[1,2,3]|in_string_array[1,2]",
+            "id" => "<string>[1,2,3]|in_string_array[1,2]",
         ];
 
         $cases = [
@@ -953,7 +953,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "(s)",
+            "error_tag" => "<string>",
             "field_path" => "id",
             "parameters" => [
                 "1,2,3"
@@ -970,7 +970,7 @@ trait TestRuleDefault
     protected function test_method_not_in_string_array()
     {
         $rule = [
-            "id" => "!(s)[1,2]|not_in_string_array[1,2,3]",
+            "id" => "!<string>[1,2]|not_in_string_array[1,2,3]",
         ];
 
         $cases = [
@@ -1020,7 +1020,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "!(s)",
+            "error_tag" => "!<string>",
             "field_path" => "id",
             "parameters" => [
                 "1,2,3"
@@ -1037,7 +1037,7 @@ trait TestRuleDefault
     protected function test_method_length_equal()
     {
         $rule = [
-            "text" => "len=[3]|length_equal[3]",
+            "text" => "length=[3]|length_equal[3]",
         ];
 
         $cases = [
@@ -1080,7 +1080,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len=",
+            "error_tag" => "length=",
             "field_path" => "text",
             "parameters" => [
                 "3"
@@ -1097,7 +1097,7 @@ trait TestRuleDefault
     protected function test_method_length_not_equal()
     {
         $rule = [
-            "text" => "len!=[3]|length_not_equal[3]",
+            "text" => "length!=[3]|length_not_equal[3]",
         ];
 
         $cases = [
@@ -1140,7 +1140,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len!=",
+            "error_tag" => "length!=",
             "field_path" => "text",
             "parameters" => [
                 "3"
@@ -1157,7 +1157,7 @@ trait TestRuleDefault
     protected function test_method_length_greater_than()
     {
         $rule = [
-            "text" => "len>[3]|length_greater_than[3]",
+            "text" => "length>[3]|length_greater_than[3]",
         ];
 
         $cases = [
@@ -1210,7 +1210,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len>",
+            "error_tag" => "length>",
             "field_path" => "text",
             "parameters" => [
                 "3"
@@ -1227,7 +1227,7 @@ trait TestRuleDefault
     protected function test_method_length_less_than()
     {
         $rule = [
-            "text" => "len<[3]|length_less_than[3]",
+            "text" => "length<[3]|length_less_than[3]",
         ];
 
         $cases = [
@@ -1280,7 +1280,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len<",
+            "error_tag" => "length<",
             "field_path" => "text",
             "parameters" => [
                 "3"
@@ -1297,7 +1297,7 @@ trait TestRuleDefault
     protected function test_method_length_greater_than_equal()
     {
         $rule = [
-            "text" => "len>=[3]|length_greater_than_equal[3]",
+            "text" => "length>=[3]|length_greater_than_equal[3]",
         ];
 
         $cases = [
@@ -1355,7 +1355,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len>=",
+            "error_tag" => "length>=",
             "field_path" => "text",
             "parameters" => [
                 "3"
@@ -1372,7 +1372,7 @@ trait TestRuleDefault
     protected function test_method_length_less_than_equal()
     {
         $rule = [
-            "text" => "len<=[3]|length_less_than_equal[3]",
+            "text" => "length<=[3]|length_less_than_equal[3]",
         ];
 
         $cases = [
@@ -1425,7 +1425,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len<=",
+            "error_tag" => "length<=",
             "field_path" => "text",
             "parameters" => [
                 "3"
@@ -1442,7 +1442,7 @@ trait TestRuleDefault
     protected function test_method_length_between()
     {
         $rule = [
-            "text" => "len<>[1,10]|length_between[1,10]",
+            "text" => "length><[1,10]|length_between[1,10]",
         ];
 
         $cases = [
@@ -1520,7 +1520,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len<>",
+            "error_tag" => "length><",
             "field_path" => "text",
             "parameters" => [
                 "1",
@@ -1538,7 +1538,7 @@ trait TestRuleDefault
     protected function test_method_length_greater_lessequal()
     {
         $rule = [
-            "text" => "len<=>[1,10]|length_greater_lessequal[1,10]",
+            "text" => "length><=[1,10]|length_greater_lessequal[1,10]",
         ];
 
         $cases = [
@@ -1616,7 +1616,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len<=>",
+            "error_tag" => "length><=",
             "field_path" => "text",
             "parameters" => [
                 "1",
@@ -1634,7 +1634,7 @@ trait TestRuleDefault
     protected function test_method_length_greaterequal_less()
     {
         $rule = [
-            "text" => "len<>=[1,10]|length_greaterequal_less[1,10]",
+            "text" => "length>=<[1,10]|length_greaterequal_less[1,10]",
         ];
 
         $cases = [
@@ -1712,7 +1712,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len<>=",
+            "error_tag" => "length>=<",
             "field_path" => "text",
             "parameters" => [
                 "1",
@@ -1730,7 +1730,7 @@ trait TestRuleDefault
     protected function test_method_length_greaterequal_lessequal()
     {
         $rule = [
-            "text" => "len<=>=[1,10]|length_greaterequal_lessequal[1,10]",
+            "text" => "length>=<=[1,10]|length_greaterequal_lessequal[1,10]",
         ];
 
         $cases = [
@@ -1813,7 +1813,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "len<=>=",
+            "error_tag" => "length>=<=",
             "field_path" => "text",
             "parameters" => [
                 "1",
@@ -2027,7 +2027,7 @@ trait TestRuleDefault
     protected function test_method_is_array()
     {
         $rule = [
-            "id" => "arr|is_array",
+            "id" => "array|is_array",
         ];
 
         $cases = [
@@ -2080,7 +2080,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "arr",
+            "error_tag" => "array",
             "field_path" => "id",
         ];
 
@@ -2153,10 +2153,10 @@ trait TestRuleDefault
         ];
     }
 
-    protected function test_method_bool_str()
+    protected function test_method_bool_string()
     {
         $rule = [
-            "id" => "bool_str",
+            "id" => "bool_string",
         ];
 
         $cases = [
@@ -2214,7 +2214,7 @@ trait TestRuleDefault
 
         $extra = [
             "method_name" => __METHOD__,
-            "error_tag" => "bool_str",
+            "error_tag" => "bool_string",
             "field_path" => "id",
         ];